% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb.R
\name{lb_auglag}
\alias{lb_auglag}
\title{Fitting Log-Binomial Models with \code{auglag} from the \pkg{alabama} Package}
\usage{
lb_auglag(x, y, start, tol = 1e-07, ceps = 1e-07, control = list(),
  control.optim = list(), implementation = c("improved", "naive"))
}
\arguments{
\item{x}{design matrix of dimension \code{n * p}.}

\item{y}{vector of observations of length \code{n}.}

\item{start}{numeric (vector) starting value (of length p).}

\item{tol}{a numeric giving the tolerance for convergence of the outer iterations 
(see \code{eps} in \code{control.outer} the from the \code{auglag} function.}

\item{ceps}{epsilon subtracted from the right hand side (\eqn{X \beta \leq 0 - ceps}).
Since the inequality \eqn{X \beta \leq 0} is only fullfilled to a certain tolerance
we have to substact epsilon to ensure \eqn{X \beta \leq 0}.}

\item{control}{passed to \code{control.outer}.}

\item{control.optim}{passed to \code{control.optim}.}

\item{implementation}{a character string choosing which implementation of the 
log-likelihood and gradient is used.}
}
\value{
the optimization result.
}
\description{
Fitting Log-Binomial Models with \code{auglag} from the \pkg{alabama} Package
}
