% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb.R
\name{lb_convex}
\alias{lb_convex}
\title{Fitting Log-Binomial Models by Convex Optimization}
\usage{
lb_convex(x, y, method = c("dense", "sparse", "weighted"), tol = 1e-08,
  ceps = 1e-07, control = list(), dry_run = FALSE, solver = "ecos")
}
\arguments{
\item{x}{design matrix of dimension \code{n * p}.}

\item{y}{vector of observations of length \code{n}.}

\item{method}{a character string giving the method used to construct the optimization problem,
possible}

\item{tol}{tolerance for the optimizer (default is \code{1e-8}).}

\item{ceps}{epsilon subtracted from the right hand side (\eqn{X \beta \leq 0 - ceps}).
Since the inequality \eqn{X \beta \leq 0} is only fullfilled to a certain tolerance
we have to substact epsilon to ensure \eqn{X \beta \leq 0}.}

\item{control}{a list containing additional arguments passed to \code{ROI_solve}.}

\item{dry_run}{a logical controling if the problem should be solved 
(default \code{dry\_run = FALSE}) or the optimization problem should be returned.}

\item{solver}{a character string selecting the solver to be used (default is \code{"ecos"}).}
}
\value{
a vector giving the estimated coefficients.
}
\description{
Fitting Log-Binomial Models by Convex Optimization
}
\examples{
d <- simdata_blizhosm2006(500, 1)
lb_convex(d$x, d$y)
}
