% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb.R
\name{lb_coptim}
\alias{lb_coptim}
\title{Fitting Log-Binomial Models with \code{constrOptim}}
\usage{
lb_coptim(x, y, start, mu = 1e-04, control = list(maxit = 500L),
  method = "BFGS", outer_iterations = 10000, outer_eps = 1e-08, ...)
}
\arguments{
\item{x}{design matrix of dimension \code{n * p}.}

\item{y}{vector of observations of length \code{n}.}

\item{start}{numeric (vector) starting value (of length p): must be in the
feasible region.}

\item{mu}{see \code{constrOptim}.}

\item{control}{passed to ``optim''.}

\item{method}{passed to ``optim''.}

\item{outer_iterations}{see \code{constrOptim}.}

\item{outer_eps}{see \code{constrOptim}.}

\item{...}{see \code{constrOptim}.}
}
\value{
a vector giving the estimated coefficients.
}
\description{
Fitting Log-Binomial Models with \code{constrOptim}
}
