% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{lb_start}
\alias{lb_start}
\title{Find a starting value}
\usage{
lb_start(y, x, method = c("lp", "simple", "poisson", "unconstrained"),
  delta = 1)
}
\arguments{
\item{y}{vector of observations of length \code{n}.}

\item{x}{design matrix of dimension \code{n * p}.}

\item{method}{a character string giving the method used to select the starting values.
The use of \code{"simple"} or \code{"lp"} is recommended.}

\item{delta}{a numeric additionaly substracted from the intercept.}
}
\value{
a vector giving the starting values.
}
\description{
Find a starting value
}
