% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_data_marschner_plus_minus}
\alias{sim_data_marschner_plus_minus}
\title{Simulate Data similar to Marschner 2018}
\usage{
sim_data_marschner_plus_minus(nobs, ncov)
}
\arguments{
\item{nobs}{number of observations.}

\item{ncov}{number of covariates.}
}
\value{
depending on rtype either a \code{list} or a \code{data.frame}.
}
\description{
This slightly modified verison allows to use more covariates.
}
\examples{
d <- sim_data_marschner_plus_minus(nobs = 200, ncov = 20L)
}
