% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simdata_blizhosm2006}
\alias{simdata_blizhosm2006}
\title{Simulate Data based on Blizzard and Hosmer 2006}
\usage{
simdata_blizhosm2006(nobs, setting, rtype = c("list", "data.frame"))
}
\arguments{
\item{nobs}{number of observations.}

\item{setting}{an integer giving the simulation setting. Allowed values are 
\eqn{\{1, 2, 3, 4, 5, 6, 7, 8\}}.}

\item{rtype}{a character giving the return type. Allowed values are \code{"list"} and \code{"data.frame"}.}
}
\value{
depending on rtype either a \code{list} or a \code{data.frame}.
}
\description{
Simulate Data based on Blizzard and Hosmer 2006
}
\examples{
d <- simdata_blizhosm2006(500, 1)
}
\references{
L. Blizzard and D. W. Hosmer. "Parameter estimation and goodness-of-fit in log binomial regression." Biometrical Journal 48.1 (2006): 5-22.
}
