% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simdata_marschner_jss_2018}
\alias{simdata_marschner_jss_2018}
\title{Simulate Data based on Marschner 2018}
\usage{
simdata_marschner_jss_2018(nobs, ncov, setting, rtype = c("list",
  "data.frame"))
}
\arguments{
\item{nobs}{number of observations.}

\item{ncov}{number of covariates.}

\item{setting}{an integer giving the simulation setting. Allowed values are \eqn{\{1, 2\}}.}

\item{rtype}{a character giving the return type. Allowed values are \code{"list"} and \code{"data.frame"}.}
}
\value{
depending on rtype either a \code{list} or a \code{data.frame}.
}
\description{
Simulate Data based on Marschner 2018
}
\examples{
d <- simdata_marschner_jss_2018(nobs = 2000, ncov = 20, setting = 1)
}
