% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simdata_rms}
\alias{simdata_rms}
\title{Simulate Data similar to an example in the \pkg{rms} Manual}
\usage{
simdata_rms(nobs, rtype = c("list", "data.frame"))
}
\arguments{
\item{nobs}{number of observations.}

\item{rtype}{a character giving the return type. Allowed values are \code{"list"} and \code{"data.frame"}.}
}
\value{
depending on rtype either a \code{list} or a \code{data.frame}.
}
\description{
Simulate Data similar to an example in the \pkg{rms} Manual
}
\examples{
d <- simdata_rms(500)
}
\references{
Frank E Harrell Jr (2019). rms: Regression Modeling Strategies. R package version 5.1-3. https://CRAN.R-project.org/package=rms
\url{https://www.rdocumentation.org/packages/rms/versions/5.1-2/topics/anova.rms}
}
