% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simdata_savuliuyasui2009}
\alias{simdata_savuliuyasui2009}
\title{Simulate Data based on Savu, Liu and Yasui 2009}
\usage{
simdata_savuliuyasui2009(nobs, setting, rtype = c("list", "data.frame"))
}
\arguments{
\item{nobs}{number of observations.}

\item{setting}{an integer giving the simulation setting. Allowed values are \eqn{\{1, 2, 3, 4, 5, 6, 7, 8\}}.}

\item{rtype}{a character giving the return type. Allowed values are \code{"list"} and \code{"data.frame"}.}
}
\value{
depending on rtype either a \code{list} or a \code{data.frame}.
}
\description{
Simulate Data based on Savu, Liu and Yasui 2009
}
\examples{
d <- simdata_savuliuyasui2009(nobs = 200, 1L)
}
\references{
Savu A, Liu Q, Yasui Y (2010) Estimation of relative risk and prevalence ratio. Statistics in Medicine 29(22):2269-2281, DOI \url{10.1002/sim.3989}
}
